+++ app/src/main/kotlin/com/waz/zclient/settings/main/list/SettingsMainListAdapter.kt
@@ -5,7 +5,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.waz.zclient.R
-import com.waz.zclient.core.lists.OnItemClickListener
+import com.waz.zclient.core.ui.list.OnItemClickListener
 import com.waz.zclient.settings.main.model.SettingsMainItem
 import kotlinx.android.synthetic.main.item_settings.view.*
